+++ src/call/hangup.rs
@@ -1,5 +1,6 @@
 //! Types for the *m.call.hangup* event.
 
+use js_int::UInt;
 use serde::{Deserialize, Serialize};
 
 room_event! {
@@ -15,7 +16,7 @@ pub struct HangupEventContent {
     pub call_id: String,
 
     /// The version of the VoIP specification this messages adheres to.
-    pub version: u64,
+    pub version: UInt,
 
     /// Optional error reason for the hangup.
     pub reason: Option<Reason>,
