+++ src/call/invite.rs
@@ -1,5 +1,6 @@
 //! Types for the *m.call.invite* event.
 
+use js_int::UInt;
 use serde::{Deserialize, Serialize};
 
 use super::SessionDescription;
@@ -18,11 +19,11 @@ pub struct InviteEventContent {
     /// The time in milliseconds that the invite is valid for. Once the invite age exceeds this
     /// value, clients should discard it. They should also no longer show the call as awaiting an
     /// answer in the UI.
-    pub lifetime: u64,
+    pub lifetime: UInt,
 
     /// The session description object. The session description type must be *offer*.
     pub offer: SessionDescription,
 
     /// The version of the VoIP specification this messages adheres to.
-    pub version: u64,
+    pub version: UInt,
 }
